requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose==3.3.0
rsa==4.9
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
